+++ Game/Plan/PlanInitial.cs
@@ -19,8 +19,7 @@ public partial class PlanInitial : Node2D
     public static bool VehiculesInit = false;
     public static Vector2 VehiculesPosition;
     public static Vehicules.Type VehiculesType;
-    private PackedScene _ambulanceScene;
-    private PackedScene _camionScene;
+    private PackedScene _vehiculeScene;
 
 
     private Vector2 _lastTile = new Vector2(0, 0);
@@ -68,8 +67,7 @@ public partial class PlanInitial : Node2D
         TileMap1 = (TileMap) GetNode(str_TileMap1);
         TileMap2 = (TileMap) GetNode(str_TileMap2);
         TileMap3 = (TileMap) GetNode(str_TileMap3);
-        _camionScene = (PackedScene) GD.Load("res://Game/Vehicules/Camion.tscn");
-        _ambulanceScene = (PackedScene) GD.Load("res://Game/Vehicules/Ambulance.tscn");
+        _vehiculeScene = (PackedScene) GD.Load("res://Game/Vehicules/Vehicules.tscn");
     }
 
     public override void _Process(float delta)
@@ -83,15 +81,9 @@ public partial class PlanInitial : Node2D
 
         if (VehiculesInit)
         {
-            if (VehiculesType == Vehicules.Type.AMBULANCE)
-            {
-                InitAmbulance(VehiculesPosition);
-            }
-            else if (VehiculesType == Vehicules.Type.CAMION)
-            {
-                InitCamion(VehiculesPosition);
-            }
-
+            Vehicules _vehicule = (Vehicules) _vehiculeScene.Instance();
+            _vehicule.Init(this, Routes.WhereIsRoute(VehiculesPosition, this), VehiculesType);
+            AddChild(_vehicule);
             VehiculesInit = false;
         }
     }
@@ -101,20 +93,6 @@ public partial class PlanInitial : Node2D
         VehiculesInit = true;
         VehiculesPosition = position;
         VehiculesType = type;
-    }
-
-    public void InitCamion(Vector2 position)
-    {
-        Camion _camion = (Camion) _camionScene.Instance();
-        _camion.Init(this, Routes.WhereIsRoute(position, this));
-        AddChild(_camion);
-    }
-
-    public void InitAmbulance(Vector2 position)
-    {
-        Ambulance _ambulance = (Ambulance) _ambulanceScene.Instance();
-        _ambulance.Init(this, Routes.WhereIsRoute(position, this));
-        AddChild(_ambulance);
     }
 
     public void SetBlock(TileMap tileMap, int x, int y, int index)
