+++ Game/Vehicules/ANimatedSpriteVehicules/Ambulance_animatedSprite.tres
@@ -1,0 +1,29 @@
+[gd_resource type="SpriteFrames" load_steps=5 format=2]
+
+[ext_resource path="res://assets/vehicules/Isometric/ambulance_SE.png" type="Texture" id=1]
+[ext_resource path="res://assets/vehicules/Isometric/ambulance_NE.png" type="Texture" id=2]
+[ext_resource path="res://assets/vehicules/Isometric/ambulance_NW.png" type="Texture" id=3]
+[ext_resource path="res://assets/vehicules/Isometric/ambulance_SW.png" type="Texture" id=4]
+
+[resource]
+animations = [ {
+"frames": [ ExtResource( 4 ) ],
+"loop": true,
+"name": "SW",
+"speed": 5.0
+}, {
+"frames": [ ExtResource( 3 ) ],
+"loop": true,
+"name": "NW",
+"speed": 5.0
+}, {
+"frames": [ ExtResource( 2 ) ],
+"loop": true,
+"name": "NE",
+"speed": 5.0
+}, {
+"frames": [ ExtResource( 1 ) ],
+"loop": true,
+"name": "SE",
+"speed": 5.0
+} ]
