+++ Game/Vehicules/ANimatedSpriteVehicules/Camion_animatedSprite.tres
@@ -12,6 +12,11 @@ animations = [ {
 "name": "SW",
 "speed": 5.0
 }, {
+"frames": [ ExtResource( 4 ) ],
+"loop": true,
+"name": "NE",
+"speed": 5.0
+}, {
 "frames": [ ExtResource( 3 ) ],
 "loop": true,
 "name": "SE",
@@ -20,10 +25,5 @@ animations = [ {
 "frames": [ ExtResource( 1 ) ],
 "loop": true,
 "name": "NW",
-"speed": 5.0
-}, {
-"frames": [ ExtResource( 4 ) ],
-"loop": true,
-"name": "NE",
 "speed": 5.0
 } ]
