--- Game/Vehicules/Ambulance.tscn
@@ -1,36 +1,0 @@
-[gd_scene load_steps=7 format=2]
-
-[ext_resource path="res://assets/vehicules/Isometric/ambulance_SE.png" type="Texture" id=1]
-[ext_resource path="res://assets/vehicules/Isometric/ambulance_NW.png" type="Texture" id=2]
-[ext_resource path="res://assets/vehicules/Isometric/ambulance_SW.png" type="Texture" id=3]
-[ext_resource path="res://assets/vehicules/Isometric/ambulance_NE.png" type="Texture" id=4]
-[ext_resource path="res://Game/Vehicules/Ambulance.cs" type="Script" id=5]
-
-[sub_resource type="SpriteFrames" id=1]
-animations = [ {
-"frames": [ ExtResource( 3 ) ],
-"loop": true,
-"name": "SW",
-"speed": 5.0
-}, {
-"frames": [ ExtResource( 1 ) ],
-"loop": true,
-"name": "SE",
-"speed": 5.0
-}, {
-"frames": [ ExtResource( 2 ) ],
-"loop": true,
-"name": "NW",
-"speed": 5.0
-}, {
-"frames": [ ExtResource( 4 ) ],
-"loop": true,
-"name": "NE",
-"speed": 5.0
-} ]
-
-[node name="Ambulance" type="AnimatedSprite"]
-scale = Vector2( 1.5, 1.5 )
-frames = SubResource( 1 )
-animation = "SW"
-script = ExtResource( 5 )
