+++ Game/Vehicules/Vehicules.tscn
@@ -1,17 +1,17 @@
 [gd_scene load_steps=4 format=2]
 
-[ext_resource path="res://Game/MainPlan/Camion_animatedSprite.tres" type="SpriteFrames" id=1]
-[ext_resource path="res://Game/Vehicules/Camion.cs" type="Script" id=2]
+[ext_resource path="res://Game/Vehicules/Vehicules.cs" type="Script" id=1]
+[ext_resource path="res://Game/Vehicules/ANimatedSpriteVehicules/Camion_animatedSprite.tres" type="SpriteFrames" id=2]
 
 [sub_resource type="RectangleShape2D" id=1]
 extents = Vector2( 128.85, 75.4471 )
 
-[node name="Camion" type="Area2D"]
+[node name="Vehicule" type="Area2D"]
 scale = Vector2( 1.5, 1.5 )
-script = ExtResource( 2 )
+script = ExtResource( 1 )
 
 [node name="AnimatedSprite" type="AnimatedSprite" parent="."]
-frames = ExtResource( 1 )
+frames = ExtResource( 2 )
 animation = "NW"
 
 [node name="CollisionShape2D" type="CollisionShape2D" parent="."]
